Move Location into its own internal package
