+++ error.go
@@ -9,6 +9,8 @@ package testdeep
 import (
 	"bytes"
 	"strings"
+
+	"github.com/maxatome/go-testdeep/internal/location"
 )
 
 // Error represents errors generated by testdeep functions.
@@ -25,7 +27,7 @@ type Error struct {
 	// Got + Expected fields
 	Summary interface{}
 	// If initialized, location of TestDeep operator originator of the error
-	Location Location
+	Location location.Location
 	// If defined, the current Error comes from this Error
 	Origin *Error
 	// If defined, points to the next Error
