+++ types.go
@@ -13,6 +13,8 @@ import (
 	"strings"
 	"testing"
 	"time"
+
+	"github.com/maxatome/go-testdeep/internal/location"
 )
 
 var (
@@ -66,7 +68,7 @@ type TestDeep interface {
 	testDeepStringer
 	Match(ctx Context, got reflect.Value) *Error
 	setLocation(int)
-	GetLocation() Location
+	GetLocation() location.Location
 	HandleInvalid() bool
 	TypeBehind() reflect.Type
 }
@@ -74,14 +76,14 @@ type TestDeep interface {
 // Base is a base type providing some methods needed by the TestDeep
 // interface.
 type Base struct {
-	location Location
+	location location.Location
 }
 
 func (t Base) _TestDeep() {}
 
 func (t *Base) setLocation(callDepth int) {
 	var ok bool
-	t.location, ok = NewLocation(callDepth)
+	t.location, ok = location.New(callDepth)
 	if !ok {
 		t.location.File = "???"
 		t.location.Line = 0
@@ -91,7 +93,7 @@ func (t *Base) setLocation(callDepth int) {
 	opDotPos := strings.LastIndex(t.location.Func, ".")
 
 	// Try to go one level upper, to check if it is a CmpXxx function
-	cmpLoc, ok := NewLocation(callDepth + 1)
+	cmpLoc, ok := location.New(callDepth + 1)
 	if ok {
 		cmpDotPos := strings.LastIndex(cmpLoc.Func, ".")
 
@@ -107,9 +109,9 @@ func (t *Base) setLocation(callDepth int) {
 	t.location.Func = t.location.Func[opDotPos+1:]
 }
 
-// GetLocation returns a copy of the Location where the TestDeep
+// GetLocation returns a copy of the location.Location where the TestDeep
 // operator has been created.
-func (t *Base) GetLocation() Location {
+func (t *Base) GetLocation() location.Location {
 	return t.location
 }
 
@@ -126,7 +128,7 @@ func (t Base) TypeBehind() reflect.Type {
 	return nil
 }
 
-// NewBase returns a new Base struct with Location set to the
+// NewBase returns a new Base struct with location.Location set to the
 // "callDepth" depth.
 func NewBase(callDepth int) (b Base) {
 	b.setLocation(callDepth)
@@ -145,7 +147,7 @@ func (t BaseOKNil) HandleInvalid() bool {
 	return true
 }
 
-// NewBaseOKNil returns a new BaseOKNil struct with Location set to
+// NewBaseOKNil returns a new BaseOKNil struct with location.Location set to
 // the "callDepth" depth.
 func NewBaseOKNil(callDepth int) (b BaseOKNil) {
 	b.setLocation(callDepth)
